Change type time difference to float

+++ db/migrate/20170218012021_create_range_times.rb
@@ -3,8 +3,8 @@ class CreateRangeTimes < ActiveRecord::Migration[5.0]
     create_table :range_times do |t|
       t.time :start_time
       t.time :end_time
-      t.integer :difference_hours
-      t.integer :difference_seconds
+      t.float :difference_hours
+      t.float :difference_seconds
       t.datetime :date
       t.references :user, foreign_key: true
